Move background image plotter into plotting module
